Move _CONFIGURATOR_CLASSES to core
