+++ src/dstack/_internal/server/services/backends/__init__.py
@@ -1,6 +1,6 @@
 import asyncio
 import heapq
-from typing import Callable, Coroutine, Dict, List, Optional, Tuple, Type, Union
+from typing import Callable, Coroutine, Dict, List, Optional, Tuple
 from uuid import UUID
 
 from sqlalchemy import delete, update
@@ -11,6 +11,7 @@ from dstack._internal.core.backends.base.configurator import (
     Configurator,
     StoredBackendRecord,
 )
+from dstack._internal.core.backends.configurators import get_configurator
 from dstack._internal.core.backends.local.backend import LocalBackend
 from dstack._internal.core.backends.models import (
     AnyBackendConfig,
@@ -36,131 +37,6 @@ from dstack._internal.utils.common import run_async
 from dstack._internal.utils.logging import get_logger
 
 logger = get_logger(__name__)
-
-_CONFIGURATOR_CLASSES: List[Type[Configurator]] = []
-
-try:
-    from dstack._internal.core.backends.aws.configurator import AWSConfigurator
-
-    _CONFIGURATOR_CLASSES.append(AWSConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.azure.configurator import AzureConfigurator
-
-    _CONFIGURATOR_CLASSES.append(AzureConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.cudo.configurator import (
-        CudoConfigurator,
-    )
-
-    _CONFIGURATOR_CLASSES.append(CudoConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.datacrunch.configurator import (
-        DataCrunchConfigurator,
-    )
-
-    _CONFIGURATOR_CLASSES.append(DataCrunchConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.gcp.configurator import GCPConfigurator
-
-    _CONFIGURATOR_CLASSES.append(GCPConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.kubernetes.configurator import (
-        KubernetesConfigurator,
-    )
-
-    _CONFIGURATOR_CLASSES.append(KubernetesConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.lambdalabs.configurator import (
-        LambdaConfigurator,
-    )
-
-    _CONFIGURATOR_CLASSES.append(LambdaConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.nebius.configurator import NebiusConfigurator
-
-    _CONFIGURATOR_CLASSES.append(NebiusConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.oci.configurator import OCIConfigurator
-
-    _CONFIGURATOR_CLASSES.append(OCIConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.runpod.configurator import RunpodConfigurator
-
-    _CONFIGURATOR_CLASSES.append(RunpodConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.tensordock.configurator import (
-        TensorDockConfigurator,
-    )
-
-    _CONFIGURATOR_CLASSES.append(TensorDockConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.vastai.configurator import VastAIConfigurator
-
-    _CONFIGURATOR_CLASSES.append(VastAIConfigurator)
-except ImportError:
-    pass
-
-try:
-    from dstack._internal.core.backends.vultr.configurator import VultrConfigurator
-
-    _CONFIGURATOR_CLASSES.append(VultrConfigurator)
-except ImportError:
-    pass
-
-
-_BACKEND_TYPE_TO_CONFIGURATOR_CLASS_MAP = {c.TYPE: c for c in _CONFIGURATOR_CLASSES}
-
-
-def register_configurator(configurator: Type[Configurator]):
-    _BACKEND_TYPE_TO_CONFIGURATOR_CLASS_MAP[configurator.TYPE] = configurator
-
-
-def get_configurator(backend_type: Union[BackendType, str]) -> Optional[Configurator]:
-    backend_type = BackendType(backend_type)
-    configurator_class = _BACKEND_TYPE_TO_CONFIGURATOR_CLASS_MAP.get(backend_type)
-    if configurator_class is None:
-        return None
-    return configurator_class()
-
-
-def list_available_backend_types() -> List[BackendType]:
-    available_backend_types = []
-    for configurator_class in _BACKEND_TYPE_TO_CONFIGURATOR_CLASS_MAP.values():
-        available_backend_types.append(configurator_class.TYPE)
-    return available_backend_types
 
 
 async def create_backend(
